fix(rendering): :bug: fix random crashes when trying to address ImGui before engine start
