+++ src/main/java/com/sevenmap/Demo.java
@@ -11,6 +11,7 @@ import java.util.Map;
 
 import com.sevenmap.ui.Engine;
 import com.sevenmap.ui.elements.GuiLayer;
+import com.sevenmap.ui.elements.GuiNode;
 import com.sevenmap.ui.elements.Item;
 import com.sevenmap.ui.gfx.Mesh;
 import com.sevenmap.ui.gfx.Vertex;
@@ -21,7 +22,7 @@ public class Demo {
     private Engine engine = new Engine();
     private Mesh surfaceMesh;
     private Item surfaceItem;
-    private GuiLayer debugStats;
+    private GuiNode debugStats;
 
     /** A hashmap containing all the keybinds in this demo */
     private HashMap<Integer, Runnable> keybinds = new HashMap<>();
@@ -62,13 +63,6 @@ public class Demo {
 
         // instanciate mesh and item and reparent the item to the RootNode
         surfaceMesh = new Mesh(vertices, indices);
-        debugStats = new GuiLayer();
-        debugStats.addLogic(() -> {
-            if (ImGui.button("Show stats")) {
-                System.out.println("button pushed");
-            }
-        });
-        debugStats.setParent(engine.getGuiRoot());
         surfaceItem = new Item(
                             new Vector3f(- TILESIZE * SURFACESIZE / 2, -0.5f, - TILESIZE * SURFACESIZE / 2), 
                             new Vector3f(0, 0, 0),
@@ -76,6 +70,7 @@ public class Demo {
                             surfaceMesh);
 
         surfaceItem.setParent(engine.getSceneRoot());
+        createGui();
         // debug
         engine.getSceneRoot().tree();
         engine.getGuiRoot().tree();
@@ -124,6 +119,16 @@ public class Demo {
         for (Map.Entry<Integer, Runnable> entry : keybinds.entrySet()) {
             engine.getWindow().onKeyDown(entry.getKey(), entry.getValue());
         }
+    }
+
+    private void createGui() {
+        // ImGui.getIO().getFonts().addFontFromFileTTF(Demo.class.getClassLoader().getResource("fonts/Raleway/static/Raleway-Regular.ttf").toString(), 20);
+        // ImGui.getIO().getFonts().addFontFromFileTTF("src/main/resources/fonts/Raleway/static/Raleway-Regular.ttf", 20);
+        debugStats = new GuiNode();
+        debugStats.addLogic(() -> {
+            ImGui.showDemoWindow();
+        });
+        debugStats.setParent(engine.getGuiRoot());
     }
 
     private float smoothstep(float x, float min, float max) {
