+++ src/main/java/com/sevenmap/ui/elements/GuiNode.java
@@ -1,14 +1,21 @@
 package com.sevenmap.ui.elements;
 
-/**
- * Unfinished, please use GuiLayer instead.
- */
+import com.sevenmap.exceptions.IncorrectChildTypeError;
+
 public class GuiNode extends Node {
+    protected Runnable logic;
     
+    /**
+     * Create a new GuiNode object with a default name.
+     */
     public GuiNode() {
         name = this.getClass().getSimpleName();
     }
 
+    /**
+     * Create a new GuiNode object with a chosen name.
+     * @param name the given name
+     */
     public GuiNode(String name) {
         this.name = name;
     }
@@ -18,6 +25,35 @@ public class GuiNode extends Node {
      * in subclass definition.
      */
     public void draw() {
-        // this method should contain your code
+        compute();
+    }
+
+    public void addLogic(Runnable logic) {
+        this.logic = logic;
+    }
+
+     /**
+     * Computer Layer logic and draw its children.
+     * <p>
+     * Override this function if you need to create your own Layer subclass (inheriting
+     * from GuiLayer),
+     * by default it iterates on the Layer's children and draws them
+     * <p/>
+     */
+    public void compute() {
+        if (logic != null) {
+            logic.run();
+        }
+        shownChildren.forEach((Node node) -> ((GuiNode) node).draw()); // do stuff
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void compatibilityCheck(Node child) {
+        if (!(child instanceof GuiNode)) {
+            throw new IncorrectChildTypeError("GuiNode element can only receive children of types GuiNode or lower.");
+        }
     }
 }
